+++ sf-logging/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "sf-logging"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+tracing = "0.1.41"
+
+[lints.rust]
+unexpected_cfgs = { level = "warn", check-cfg = ['cfg(coverage)'] }
