+++ sf-server/src/builder.rs
@@ -85,6 +85,7 @@ where
 }
 
 #[cfg(test)]
+#[cfg_attr(coverage_nightly, coverage(off))]
 mod tests {
     use super::*;
     use axum::{response::IntoResponse, routing::get};
