+++ sf-server/src/error.rs
@@ -14,4 +14,15 @@ pub enum Error {
     /// Error when trying to add a peer that already exists
     #[error("Peer already exists: {0}")]
     PeerAlreadyExists(String),
+
+    /// The peer‑specific mpsc channel is closed, so the message could
+    /// not be delivered.
+    #[error("Send error: peer receiver has been dropped")]
+    SendChannelClosed,
+}
+
+impl<T> From<tokio::sync::mpsc::error::SendError<T>> for Error {
+    fn from(_: tokio::sync::mpsc::error::SendError<T>) -> Self {
+        Error::SendChannelClosed
+    }
 }
