+++ sf-server/src/socket_metadata.rs
