--- sf-test/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "sf-test"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-sf-metrics = { path = "../sf-metrics" }
-sf-logging = { path = "../sf-logging" }
-sf-protocol = { path = "../sf-protocol" }
-axum = { version = "0.8.3", features = ["ws", "macros"] }
-tokio = { version = "1.44.2", features = ["macros", "rt-multi-thread"] }
-tracing = "0.1.41"
-tracing-subscriber = "0.3.19"
-tokio-tungstenite = "0.26.2"
-serde = { version = "1.0.219", features = ["derive"] }
-tower = "0.5.2"
-http-body-util = "0.1.3"
-axum-extra = { version = "0.10.1", features = ["query"] }
-serde_json = "1.0.140"
-thiserror = "2.0.12"
-clap = { version = "4.5.37", features = ["env", "derive"] }
-serial_test = "3.2.0"
-tower-http = { version = "0.6", features = ["cors", "trace"] }
-dashmap = "6.1.0"
-futures = "0.3.31"
-
-[dev-dependencies]
-libc = "0.2.172"
-tracing-test = "0.2"
-
-[profile.dev]
-incremental = true
-
-[lints.rust.unexpected_cfgs]
-level = "deny"
-check-cfg = ['cfg(coverage)', 'cfg(coverage_nightly)']
