--- sf-test/src/peer_id.rs
@@ -1,3 +1,0 @@
-use std::sync::Arc;
-
-pub type PeerID = Arc<String>;
